+++ docs/changelog.rst
@@ -5,6 +5,8 @@ Changelog
 * :feature:`7092` Users of metamask swaps will now see them properly decoded in the history view and have them taken into account during accounting.
 * :feature:`7051` Users will be able to add a custom price directly from the asset overview page.
 * :feature:`4841` The rotki data directory is now more organized.
+
+* :release:`1.31.2 <2023-12-22>`
 * :bug:`-` Importing timestamp in CSV in scientific format should now work without errors.
 * :bug:`7123` Adding an EVM EOA address that has only withdrawals/blocks activity will no longer fail.
 * :bug:`7082` Now disabling sync for an exchange instance won't prevent other instances in the same exchange from querying new trades.
@@ -20,6 +22,7 @@ Changelog
 * :bug:`-` Withdrawal events for ethereum staking will now respect accounting rules.
 * :bug:`7059` Fix the issue where adding an account via MetaMask is not working.
 * :bug:`7057` Fix the issue where the app restarts when the file upload button is clicked.
+* :bug:`-` rotki should no longer try mainnet uniswap onchain price queries for other evm tokens and errors should be handled better
 
 * :release:`1.31.1 <2023-12-06>`
 * :bug:`-` Fix the bug that causes the error screen to appear strangely in dark mode.